requirements.txt
gradio==3.39.0
Pillow==9.5.0
python-dotenv==1.0.0
transformers==4.35.2
#torch==2.1.0+cu118
torch==2.1.0
diffusers[torch]==0.27.2
psutil

boto3~=1.34.79
numpy~=1.26.4
botocore~=1.34.79